+++ src/config.js
@@ -1,6 +1,10 @@
+
 module.exports = {
-  NODE_SEP: ';',
-  STRING_SEP: ',',
-  TERMINAL_PREFIX: '!',
-  BASE: 36,
-}
+  L1 : 32 * 32,
+  L2 : 32,
+  /**
+      The width of each unit of the encoding, in bits. Here we use 6, for base-64
+      encoding.
+   */
+  W : 6
+};
