--- src/pack/index.js
@@ -1,9 +1,0 @@
-'use strict';
-const Trie = require('./trie');
-
-//turn an array into a compressed string
-const pack = function(arr) {
-  let t = new Trie(arr)
-  return t.pack();
-}
-module.exports = pack
