--- src/unpack/index.js
@@ -1,7 +1,0 @@
-'use strict';
-const Ptrie = require('./ptrie');
-
-const unpack = (str) => {
-  return new Ptrie(str)
-}
-module.exports = unpack
